update for change in test harness callback protocol + refactoring
